Initial interface
The user can select the criteria they desire to search the database
with through a series of drop down lists for year, month, day, hour,
and minute for both start and end dates/times. User can also search by
a specific tag. Results are displayed in a sortable table with search
functionality for the whole table and individual columns as well as
pagination. From the table of results the user is able to save the
results as either a pdf or excel file, copy the results to the
clipboard, or display them in a printer friendly format.