Local library dependancies
Including the timepicker and multiselect jQuery libraries that are
needed for the interface. The timepicker library was developed by Trent
Richardson and made available via the MIT License for both personal and
commercial use. The multiselect library was developed by Eric Hynds and
made available via the MIT License and GPL License